Group reads.
Before, when writing a tiff file, each tile or strip or other memory
chunk from a data block was read from the source as a separate call.  It
turns out that this read has substantial overhead per call in python
(only a millisecond or so, but for a large tiled image this ends up
being the dominant time cost).  By grouping together reads when there is
no gap in the source data, the number of read calls can be greatly
reduced; the results are identical and the overhead is avoided.

In one benchmark of a 220 Mb file, rewriting the file when from taking
~30 seconds to ~3 seconds.